Move Zxing implementation to separate file
